Add --provider eodhd CLI flag; fix object-dtype crash on empty price columns
run.py now supports python -m fasflocken.run {demo,grid,full} --provider
eodhd (plus --cache-dir/--max-workers passthrough), so the CLI can drive
EODHDProvider directly instead of only SyntheticUniverseProvider.

Found via a real-data validation run: EODHDProvider._fetch_eod's
fallback for an unresolvable ticker built an untyped empty DataFrame
(pd.DataFrame(columns=["close"])), which defaults to object dtype. Once
concatenated with other tickers' float64 columns into one price panel,
np.log() crashed across the whole DataFrame, not just that column
("loop of ufunc does not support argument 0 of type float which has no
callable log method"). Fixed at the source (explicit float dtype) and
defensively at every np.log(prices)/np.log(etf_prices) call site
(pipeline.py, backtest.py, stats.py) so any provider returning a
non-numeric column can't take down the whole signal pipeline.

+++ fasflocken/backtest.py
@@ -112,7 +112,7 @@ def run_backtest(
         z_panel = build_z_panel(sector_signals, twin=False)
 
     etfs = [SECTOR_TO_ETF[s] for s in sectors]
-    etf_prices = provider.sector_etf_prices(start, end, etfs).reindex(index=daily_dates)
+    etf_prices = provider.sector_etf_prices(start, end, etfs).reindex(index=daily_dates).astype(float)
     etf_log_returns = np.log(etf_prices).diff()
 
     signal_dates = z_panel.index[(z_panel.index >= pd.Timestamp(start)) & (z_panel.index <= pd.Timestamp(end))]
